chore: rename info to utils
