+++ mylar/queues/__init__.py
@@ -1,12 +1,12 @@
 """Queue worker package."""
 
 from .ddl import ddl_cleanup, ddl_downloader
-from .info import QueueInfo, queue_info
 from .jd2 import jd2_queue_monitor
 from .nzb import cdh_monitor, nzb_monitor
 from .postprocess import postprocess_main
 from .search import search_queue
 from .torrent import worker_main
+from .utils import QueueInfo, queue_info
 
 __all__ = [
     "ddl_downloader",
